+++ src/Auth/CacheManager.php
@@ -6,7 +6,6 @@ namespace DiegoVasconcelos\AuthCache\Auth;
 
 use DiegoVasconcelos\AuthCache\Contracts\Cache\CacheConfigurationInterface;
 use DiegoVasconcelos\AuthCache\Contracts\Cache\CacheInterface;
-use DiegoVasconcelos\AuthCache\Contracts\Cache\CacheKeyGeneratorInterface;
 use Illuminate\Contracts\Cache\Repository;
 
 class CacheManager implements CacheInterface
@@ -14,12 +13,11 @@ class CacheManager implements CacheInterface
     public function __construct(
         private Repository $cache,
         private CacheConfigurationInterface $configuration,
-        private CacheKeyGeneratorInterface $keyGenerator
     ) {}
 
     public function remember(string $key, int|\DateTimeInterface $ttl, callable $callback): mixed
     {
-        if (! $this->isEnabled()) {
+        if (! $this->configuration->isEnabled()) {
             return $callback();
         }
 
@@ -29,20 +27,5 @@ class CacheManager implements CacheInterface
     public function forget(string $key): void
     {
         $this->cache->forget($key);
-    }
-
-    public function generateKey(string $model, mixed $identifier): string
-    {
-        return $this->keyGenerator->generate($model, $identifier);
-    }
-
-    public function isEnabled(): bool
-    {
-        return $this->configuration->isEnabled();
-    }
-
-    public function getTtl(): int
-    {
-        return $this->configuration->getTtl();
     }
 }